Impl navigate to MaterialDialog screen
