+++ app/src/main/java/com/numero/material_gallery/components/dialog/MaterialDialogFragment.kt
@@ -1,0 +1,64 @@
+package com.numero.material_gallery.components.dialog
+
+import android.content.Context
+import android.content.Intent
+import android.os.Bundle
+import android.view.ContextThemeWrapper
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import androidx.navigation.fragment.findNavController
+import com.google.android.material.dialog.MaterialAlertDialogBuilder
+import com.numero.material_gallery.R
+import com.numero.material_gallery.repository.ConfigRepository
+import kotlinx.android.synthetic.main.fragment_material_dialog.*
+import org.koin.android.ext.android.inject
+
+class MaterialDialogFragment : Fragment() {
+
+    private val configRepository by inject<ConfigRepository>()
+
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
+        return LayoutInflater
+                .from(ContextThemeWrapper(context, configRepository.themeRes))
+                .inflate(R.layout.fragment_material_dialog, container, false)
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        toolbar.apply {
+            setNavigationOnClickListener {
+                findNavController().popBackStack()
+            }
+            inflateMenu(R.menu.menu_common)
+            setOnMenuItemClickListener {
+                when (it.itemId) {
+                    R.id.action_info -> {
+                        findNavController().navigate(R.id.action_show_ThemeInfoDialog)
+                        true
+                    }
+                    else -> false
+                }
+            }
+        }
+
+        showDialogButton.setOnClickListener {
+            showMaterialDialog()
+        }
+    }
+
+    private fun showMaterialDialog() {
+        MaterialAlertDialogBuilder(requireActivity())
+                .setTitle("Title")
+                .setMessage("Message")
+                .setPositiveButton("OK", null)
+                .setNegativeButton("Cancel", null)
+                .show()
+    }
+
+    companion object {
+        fun createIntent(context: Context): Intent = Intent(context, MaterialDialogFragment::class.java)
+    }
+}
