+++ app/src/main/res/navigation/nav_graph.xml
@@ -20,6 +20,9 @@
             android:id="@+id/action_ComponentList_to_ShapeableImage"
             app:destination="@id/ShapeableImageScreen" />
         <action
+            android:id="@+id/action_ComponentList_to_MaterialDialog"
+            app:destination="@id/MaterialDialogScreen" />
+        <action
             android:id="@+id/action_ComponentList_to_NavigationDrawer"
             app:destination="@id/NavigationDrawerScreen" />
         <action
@@ -57,6 +60,11 @@
         android:id="@+id/ShapeableImageScreen"
         android:name="com.numero.material_gallery.components.image.ShapeableImageViewFragment"
         tools:layout="@layout/fragment_shapeable_image_view" />
+
+    <fragment
+        android:id="@+id/MaterialDialogScreen"
+        android:name="com.numero.material_gallery.components.dialog.MaterialDialogFragment"
+        tools:layout="@layout/fragment_material_dialog" />
 
     <fragment
         android:id="@+id/NavigationDrawerScreen"
